重构: 拆分 Group 分组方法至 Record.Group.cs
将 Record.Query.cs 中的 Group 分组方法迁移到新建的 Record.Group.cs 文件，实现未变，仅做结构调整。同时移除 RecordGroupTests.cs 中所有 Group 相关测试用例。
